GG-165 getseq en database
