+++ controllers/language_ctrl.py
@@ -1,31 +1,16 @@
 from flask import render_template, request, jsonify, redirect, url_for
 from pymongo.collection import Collection
 from models.language import Language
-
+from database import get_next_sequence_value as get_next_sequence_value
 class LanguageCtrl:
     @staticmethod
     def render_template(db: Collection):
         languagesReceived = db.find()
         return render_template('DB_Language.html', languages=languagesReceived)
 
-    @staticmethod        
-    def get_next_sequence_value(db: Collection, sequence_name):        
-        counter = db.find_one({"_id": sequence_name})
-
-        if counter is None:        
-            db.insert_one({"_id": sequence_name, "sequence_value": 1})
-            return 1
-        
-        updated_counter = db.find_one_and_update(
-            {"_id": sequence_name},
-            {"$inc": {"sequence_value": 1}},
-            return_document=True
-        )
-        return updated_counter["sequence_value"]
-
     @staticmethod
     def addLanguage(db: Collection):
-        idLanguage = LanguageCtrl.get_next_sequence_value(db,"idLanguage")
+        idLanguage = get_next_sequence_value(db,"idLanguage")
         name = request.form['name']
         if idLanguage and name:
             language = Language(idLanguage,name)
